+++ app/Http/Controllers/api/v1/ReviewController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\v1\Review\UpdateReviewRequest;
 use App\Http\Resources\v1\Review\ReviewCollection;
 use App\Http\Resources\v1\Review\ReviewResource;
 use App\Models\Review;
+use Illuminate\Support\Facades\Auth;
 use Spatie\QueryBuilder\QueryBuilder;
 
 class ReviewController extends Controller
@@ -41,16 +42,33 @@ class ReviewController extends Controller
 
     public function update(UpdateReviewRequest $request, Review $review)
     {
-        $data = $request->validated();
-        $review->update($data);
+        $userId = Auth::user()->id;
 
-        return new ReviewResource($review);
+        if ($review->user_id == $userId) {
+            $data = $request->validated();
+            $review->update($data);
+
+            return new ReviewResource($review);
+        }
+
+        return response()->json([
+            'message' => "Доступ запрещен"
+        ], 403);
+
     }
 
     public function destroy(Review $review)
     {
-        $review->delete();
+        $userId = Auth::user()->id;
 
-        return response()->noContent();
+        if ($review->user_id == $userId) {
+            $review->delete();
+
+            return response()->noContent();
+        }
+        return response()->json([
+            'message' => "Доступ запрещен"
+        ], 403);
+
     }
 }
